Restructure the tests quite a bit
