+++ contracts/CodexTitleCore.sol
@@ -11,11 +11,11 @@ import "./CodexTitleFees.sol";
 contract CodexTitleCore is CodexTitleMetadata, CodexTitleFees {
 
   /**
-    * @dev This event is emitted when a new token is minted and allows providers
-    *  to discern which Minted events came from transactions they submitted vs
-    *  transactions submitted by other platforms, as well as providing information
-    *  about what metadata record the newly minted token should be associated with
-    */
+   * @dev This event is emitted when a new token is minted and allows providers
+   *  to discern which Minted events came from transactions they submitted vs
+   *  transactions submitted by other platforms, as well as providing information
+   *  about what metadata record the newly minted token should be associated with
+   */
   event Minted(uint256 _tokenId, string _providerId, string _providerMetadataId);
 
   /**
