+++ contracts/ERC721/ERC721BasicToken.sol
@@ -188,17 +188,17 @@ contract ERC721BasicToken is ERC721Basic, ERC165 {
   }
 
   /**
-    * @dev Safely transfers the ownership of a given token ID to another address
-    * @dev If the target address is a contract, it must implement `onERC721Received`,
-    *  which is called upon a safe transfer, and return the magic value
-    *  `bytes4(keccak256("onERC721Received(address,uint256,bytes)"))`; otherwise,
-    *  the transfer is reverted.
-    * @dev Requires the msg sender to be the owner, approved, or operator
-    * @param _from current owner of the token
-    * @param _to address to receive the ownership of the given token ID
-    * @param _tokenId uint256 ID of the token to be transferred
-    * @param _data bytes data to send along with a safe transfer check
-    */
+   * @dev Safely transfers the ownership of a given token ID to another address
+   * @dev If the target address is a contract, it must implement `onERC721Received`,
+   *  which is called upon a safe transfer, and return the magic value
+   *  `bytes4(keccak256("onERC721Received(address,uint256,bytes)"))`; otherwise,
+   *  the transfer is reverted.
+   * @dev Requires the msg sender to be the owner, approved, or operator
+   * @param _from current owner of the token
+   * @param _to address to receive the ownership of the given token ID
+   * @param _tokenId uint256 ID of the token to be transferred
+   * @param _data bytes data to send along with a safe transfer check
+   */
   function safeTransferFrom(
     address _from,
     address _to,
