+++ contracts/mocks/ERC20/StandardToken.sol
@@ -15,7 +15,6 @@ contract StandardToken is ERC20, BasicToken {
 
   mapping (address => mapping (address => uint256)) internal allowed;
 
-
   /**
    * @dev Transfer tokens from one address to another
    * @param _from address The address which you want to send tokens from
